Add search_replace.py
The original use case for this script was when the Minecraft server
changed worlds. This meant introducing multiple worlds with strings
like 'world' and now 'world_new' into thousands of player data
files. This script solves this problem by searching and replacing
all the strings to make them consistent.

This will take many text files at once and replace the needed
string without having to run it individually on every file. The
only shortcoming of this script is that there are multiple strings
which will need to be replaced, which means a lot of backup files
will be created. There's probably a more elegant way of taking
multiple strings and replacing them at once, instead of having to
re-run the script multiple times.