Add new role: teleirc/default
This commit expands on PR #17 to automate more of the Teleirc bridge
bots I am using into my config management. This commit targets all the
bridge bots currently running on irc-lug.rit.edu.

Once I receive root access to the machine, I intend to implement this
role immediately on the host using the new playbook added in this
commit. After the role is tested, the PR will be merged.