:mage: nginx: Refactor into modular task files
This commit breaks down the nginx role into more modular, bite-sized
pieces. The one long file was getting irritating to scroll through, so
this makes it easier for me to work with.

Also clears away some cruft, where a templated file was stored in
`files/` instead of `templates/`.

Also also, this adds an `nginx` tag to the `horizon.yml` playbook. For
some reason, I set it to `php-fpm` only by mistake.

Signed-off-by: Justin W. Flory <git@jwf.io>