:up: php-fpm: Update from PHP 7.1 to PHP 7.3
I bumped the version of PHP on the host system to PHP 7.3. Along with
that, there were a number of differences in the default system config
files. This commit reconciles those new defaults with my custom changes.

In the future, drop-in overrides are a more convenient way of
maintaining my custom changes. Guess I should explore that.

Signed-off-by: Justin W. Flory <git@jwf.io>