:customs: sites/horizon: Transition into an extension to nginx role
This commit generifies this role a bit. So, you can add this site to a
site that uses the `nginx` role, and it will become the "default" pages
served. It replaces the default NGINX index pages.

I thought it was cute! And it also enables similar examples for other
sites.

Signed-off-by: Justin W. Flory <git@jwf.io>