:loudspeaker: xenforo: Initial commit
This commit is a rework of the `sites/crystalcraftmc.com` role. This
makes it a more generic `xenforo` role. It assumes you have a running
copy of XenForo v1.x.x to go along with it. So, it is not totally
automated but it is good for maintaining an existing installation.

Signed-off-by: Justin W. Flory <git@jwf.io>